feat: 增加商品类，收藏类接口查询

+++ src/module/commit/commit.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CommitEntity } from 'src/entities/commit.entity';
@@ -6,6 +6,8 @@ import { CommitModel } from 'src/model/commit-model';
 import { PageBean } from 'src/utils/page-bean';
 import { CollectEntity } from 'src/entities/collect.entity';
 import { CollectModel } from 'src/model/collect-model';
+import ErrorInfoMessage from 'src/common/event-key/error-info.message';
+import { GoodsEntity } from 'src/entities/goods.entity';
 
 @Injectable()
 export class CommitService {
@@ -51,12 +53,40 @@ export class CommitService {
       .where('CommitEntity.userId = :userId', { userId }).skip(skip).take(size).getMany();
     return pageBean;
   }
+  /**
+   * 保存收藏
+   * @param collectModel
+   */
   async saveCollect(collectModel: CollectModel) {
     collectModel.cTime = new Date().getTime();
     return await this.collectRepository.save(collectModel);
   }
+  /**
+   * 取消收藏
+   * @param id
+   */
   async cancelCollect(id: string) {
-    await this.collectRepository.delete(id);
-    return true;
+    const data = await this.collectRepository.findOne({ id });
+    if (data) {
+      await this.collectRepository.delete(id);
+      return true;
+    } else {
+      throw new HttpException(ErrorInfoMessage.PARAMS_ERROR, HttpStatus.BAD_REQUEST);
+    }
   }
+  /**
+   * 根据用户的id获取用户收藏信息
+   * @param id
+   */
+  async getCollectInfoByUserId(userId: string) {
+    if (!userId) {
+      throw new HttpException(ErrorInfoMessage.PARAMS_NOT_FOUND, HttpStatus.BAD_REQUEST);
+    }
+    // 关联查询
+    return await this.collectRepository.createQueryBuilder('collect')
+      .leftJoinAndMapOne('collect.targetId', GoodsEntity, 'goods', 'collect.targetId = goods.id')
+      .where('collect.userId = :userId', {userId})
+      .getMany();
+  }
+
 }
